agentic/requirements.txt
# Core dependencies for Agentic API
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
psycopg>=3.1.0
pgvector>=0.2.0

# AI/ML
phidata>=2.0.0
ollama>=0.1.0
pypdf>=3.0.0

# File handling
aiofiles>=23.0.0
python-multipart>=0.0.6

# HTTP client
requests>=2.31.0
httpx>=0.24.0